In the gui, removed the test button. Added a new button to make a new password with imputing credientals. Also can generate a new password.
